requirements.txt
# Core packages
numpy==1.24.3
scipy==1.10.1
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# Deep learning
torch==2.0.1
monai==1.2.0
nibabel==5.1.0
torchvision==0.15.2

# Web application
flask==2.3.3
flask-wtf==1.1.1
werkzeug==2.3.7

# Statistical analysis
statsmodels==0.14.0
pingouin==0.5.3

# Utilities
tqdm==4.65.0
PyYAML==6.0.1
joblib==1.3.2